+++ src/main/java/com/vata/join/service/UserService.java
@@ -1,0 +1,70 @@
+package com.vata.join.service;
+
+import com.vata.join.domain.User;
+import com.vata.join.dto.SignupRequest;
+import com.vata.join.repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final AuthenticationManager authenticationManager;
+
+    @Transactional // signup 메서드에 @Transactional 적용
+    public void signup(SignupRequest signupRequest) {
+        // 중복 검사
+        if (userRepository.findByEmail(signupRequest.email()).isPresent()) {
+            throw new IllegalArgumentException("이미 사용 중인 이메일 주소입니다.");
+        }
+
+        User user = new User();
+        user.setPassword(passwordEncoder.encode(signupRequest.password()));
+        // 비밀번호 암호화하여 User 객체에 저장
+        user.setName(signupRequest.name());
+        user.setEmail(signupRequest.email());
+
+        userRepository.save(user);  //JpaRepository 로부터 상속
+    }
+
+    /*
+    로그인 처리 메서드
+        - request 파라미터 : HTTP 요청 객체, 세션 생성 및 관리하는 데 사용
+        - 사용자 조회 : 데이터베이스 -> findByEmail -> 해당 user 객체 생성
+        - 비밀번호 확인 : 데이터베이스 pw와 파라미터 pw 비교 -> session 생성 및 정보 저장
+     */
+    public void login(String email, String password, HttpServletRequest request) {
+        try {
+            // 1. 인증 토큰 생성 (사용자 입력 정보)
+            UsernamePasswordAuthenticationToken authToken =
+                    new UsernamePasswordAuthenticationToken(email, password);
+
+            // 2. AuthenticationManager를 통해 인증 시도
+            //    loadUserByUsername 호출 -> 비밀번호 검증 (PasswordEncoder)
+            Authentication authentication = authenticationManager.authenticate(authToken);
+
+            // 3. SecurityContextHolder에 인증 정보 저장 (세션에 저장될 내용)
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+
+            // 4. 세션에 SecurityContext 저장 (Spring Security가 자동으로 세션을 관리)
+            //    기존 수동 세션 관리를 이 부분으로 대체
+            HttpSession session = request.getSession(true);
+            session.setAttribute(HttpSessionSecurityContextRepository.SPRING_SECURITY_CONTEXT_KEY,
+                    SecurityContextHolder.getContext());
+        } catch (BadCredentialsException e) {
+            throw new IllegalArgumentException("이메일 또는 비밀번호가 잘못되었습니다.");
+        }
+    }
+}
